Second commit, sorry once again for lateness
Basically, changed a huge bunch of shit. Actually made the buttons work,
changed the buttons, made the drawing functions better and less laggy.
Fixed integer rounding issues in BrownianParticle.updatePos

+++ Stochastic process simulation/src/stochasticSimulation/Simulable.java
@@ -8,9 +8,9 @@ public interface Simulable {
 	
 	/**Updates the current position of the particle
 	 * 
-	 * @param dt: timestep
+	 * @param dt: timestep in milliseconds
 	 */
-	public void updatePos(double dt);
+	public void updatePos(int dt);
 	
 	
 	/**returns current position
